refactor: dex-pool을 dex-simulation으로 이름 변경
- dex-pool/ → dex-simulation/ 폴더명 변경
- DexPoolModule → DexSimulationModule 클래스명 변경
- 블록체인 없이 순수 DEX 작동 원리만 시뮬레이션하는 목적을 명확히 반영
- 모든 import 경로 업데이트
- 소스 코드 및 테스트 코드 모두 업데이트
- 빌드 및 테스트 통과 확인 (16 suites, 252 tests)
