+++ src/app.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { EventEmitterModule } from '@nestjs/event-emitter';
-import { DexPoolModule } from './dex-pool/dex-pool.module';
+import { DexSimulationModule } from './dex-simulation/dex-simulation.module';
 import { SharedModule } from './shared/shared.module';
 
 @Module({
-  imports: [EventEmitterModule.forRoot(), SharedModule, DexPoolModule],
+  imports: [EventEmitterModule.forRoot(), SharedModule, DexSimulationModule],
   controllers: [],
   providers: [],
 })
