+++ src/dex-simulation/lp/lp.module.ts
