+++ src/dex-simulation/lp/types/lp.interface.ts
