+++ src/dex-simulation/lp/types/market.interface.ts
