+++ src/dex-simulation/market/types/market.interface.ts
