+++ src/dex-simulation/trader/trader.controllers.ts
