+++ src/dex-simulation/trader/trader.module.ts
