+++ src/dex-simulation/trader/trader.service.ts
