+++ src/dex-simulation/trader/types/arbitrage.interface.ts
