+++ src/dex-simulation/trader/types/trade.interface.ts
