+++ test/src/dex-simulation/trader/trader.service.spec.ts
@@ -1,11 +1,11 @@
 import { EventEmitter2 } from '@nestjs/event-emitter';
 import { Test, TestingModule } from '@nestjs/testing';
-import { LpService } from 'src/dex-pool/lp/lp.service';
-import { Pool } from 'src/dex-pool/lp/types/lp.interface';
-import { MarketService } from 'src/dex-pool/market/market.service';
-import { MarketPrice } from 'src/dex-pool/market/types/market.interface';
-import { TraderService } from 'src/dex-pool/trader/trader.service';
-import { ArbitrageOpportunity } from 'src/dex-pool/trader/types/arbitrage.interface';
+import { LpService } from 'src/dex-simulation/lp/lp.service';
+import { Pool } from 'src/dex-simulation/lp/types/lp.interface';
+import { MarketService } from 'src/dex-simulation/market/market.service';
+import { MarketPrice } from 'src/dex-simulation/market/types/market.interface';
+import { TraderService } from 'src/dex-simulation/trader/trader.service';
+import { ArbitrageOpportunity } from 'src/dex-simulation/trader/types/arbitrage.interface';
 
 describe('TraderService', () => {
   let service: TraderService;
